+++ docs/conf.py
@@ -23,7 +23,7 @@ copyright = '2020, sahahn'
 author = 'sahahn'
 
 # The full version, including alpha/beta/rc tags
-release = '1.3.5'
+release = '1.3.6'
 
 
 # -- General configuration ---------------------------------------------------